+++ src/app/app.module.ts
@@ -20,17 +20,18 @@ import { ContactUsComponent } from './contact-us/contact-us.component';
 import { FooterComponent } from './footer/footer.component';
 import { LoginSignupComponent } from './login-signup/login-signup.component';
 
-import { Gallery2Component } from './gallery2/gallery2.component';
-import { EggComponent } from './egg/egg.component';
-import { SpinachComponent } from './spinach/spinach.component';
-import { GarlicbreadComponent } from './garlicbread/garlicbread.component';
-import { ShrimpComponent } from './shrimp/shrimp.component';
-import { MisoComponent } from './miso/miso.component';
-import { TantanmenComponent } from './tantanmen/tantanmen.component';
-import { CurryComponent } from './curry/curry.component';
-import { KimchiComponent } from './kimchi/kimchi.component';
+// import { Gallery2Component } from './gallery2/gallery2.component';
+import { EggComponent } from './att/egg/egg.component';
+import { SpinachComponent } from './att/spinach/spinach.component';
+import { GarlicbreadComponent } from './att/garlicbread/garlicbread.component';
+import { ShrimpComponent } from './att/shrimp/shrimp.component';
+import { MisoComponent } from './att/miso/miso.component';
+import { TantanmenComponent } from './att/tantanmen/tantanmen.component';
+import { CurryComponent } from './att/curry/curry.component';
+import { KimchiComponent } from './att/kimchi/kimchi.component';
 import { FormsModule } from '@angular/forms';
 import { CommentComponent } from './comment/comment.component';
+
 
 @NgModule({
   declarations: [
@@ -46,7 +47,6 @@ import { CommentComponent } from './comment/comment.component';
     FooterComponent,
     LoginSignupComponent,
     CommentComponent,
-    Gallery2Component,
     EggComponent,
     SpinachComponent,
     GarlicbreadComponent,
@@ -54,7 +54,8 @@ import { CommentComponent } from './comment/comment.component';
     MisoComponent,
     TantanmenComponent,
     CurryComponent,
-    KimchiComponent
+    KimchiComponent,
+  
   ],
   imports: [
     BrowserModule,
