+++ src/app/att/curry/curry.component.ts
