+++ src/app/att/egg/egg.component.css
