+++ src/app/att/egg/egg.component.html
