+++ src/app/att/egg/egg.component.spec.ts
