+++ src/app/att/egg/egg.component.ts
