+++ src/app/att/kimchi/kimchi.component.ts
