+++ src/app/att/shrimp/shrimp.component.ts
