+++ src/app/att/spinach/spinach.component.ts
