+++ src/app/gallery/gallery.component.html
@@ -34,7 +34,7 @@
 
         <div class="Gallery">
             <div class="zoom"><img src="../assets/COOK BOOK MY EDITS/7.png"></div>
-            <div class="Desc"><a href="curry">Curry Omurice</a></div>
+            <div class="Desc"><a href="indcurr">Curry Omurice</a></div>
         </div>
 
         <div class="Gallery">
