--- src/app/gallery2/gallery2.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-gallery2',
-  templateUrl: './gallery2.component.html',
-  styleUrls: ['./gallery2.component.css']
-})
-export class Gallery2Component implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
